flask/requirements.txt
flask
spotipy
sqlalchemy
sqlalchemy.orm
psycopg2-binary
python-dotenv
gunicorn